Refactor: * Rename AddProofDTO to AddProof and moved it to * fix editing proof
