--- src/main/java/com/provedcode/talent/model/dto/AddProofDTO.java
@@ -1,10 +1,0 @@
-package com.provedcode.talent.model.dto;
-
-import org.hibernate.validator.constraints.URL;
-
-public record AddProofDTO(
-        @URL
-        String link,
-        String text
-) {
-}
